style: moves item_type to another path
