+++ test/src/week_2/class_2/stack/stack_test.cpp
@@ -1,7 +1,6 @@
 #include <stdexcept>
 #include "../../../../resources/test_base.h"
 #include "../../../../../src/week_2/class_2/stack/stack.cpp"
-#include <stdexcept>
 #include <cstddef>
 
 bool isStackEmpty(ItemType *structure) {
